--- usecases/api_key.js
@@ -1,24 +1,0 @@
-import { Lamatic } from "lamatic";
-
-const lamatic = new Lamatic({
-  apiKey: "your-api-key",
-  projectId: "your-project-id",
-  endpoint: "your-endpoint",
-});
-
-
-async function main() {
-  const flowId = "your-flow-id";
-  // sample payload
-  const payload = {
-    prompt: "hey, how are you?"
-  }
-  try {
-    const response = await lamatic.executeFlow(flowId, payload);
-    console.log(response);
-  } catch (error) {
-    console.error(error);
-  }
-}
-
-main();
