--- usecases/next.js/.env.local
@@ -1,4 +1,0 @@
-LAMATIC_PROJECT_ENDPOINT=your-project-endpoint
-LAMATIC_FLOW_ID=your-flow-id
-LAMATIC_PROJECT_ID=your-project-id
-LAMATIC_PROJECT_API_KEY=your-project-api-key
